requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
abseil-cpp=20211102.0=h27087fc_1
accelerate=0.33.0=pyhd8ed1ab_0
aiohttp=3.8.1=py39hb9d737c_1
aiosignal=1.3.1=pyhd8ed1ab_0
annotated-types=0.6.0=py39h06a4308_0
antlr4-python3-runtime=4.8=pypi_0
anyio=4.2.0=py39h06a4308_0
arrow=1.2.3=py39h06a4308_1
arrow-cpp=14.0.2=h374c478_1
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.10.55=h721c034_0
binaryornot=0.4.4=pyhd3eb1b0_1
blas=1.0=mkl
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.3.7=py39ha9d4c09_0
brotli-python=1.0.9=py39h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.7.4=hbcca054_0
certifi=2024.7.4=pyhd8ed1ab_0
chardet=4.0.0=py39h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py39h06a4308_0
cloudpickle=3.0.0=pypi_0
cmake=3.30.2=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
cookiecutter=2.6.0=py39h06a4308_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.5=3
datasets=2.19.1=py39h06a4308_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.8=pyhd8ed1ab_0
diskcache=5.6.3=pypi_0
distro=1.9.0=py39h06a4308_0
editdistance=0.8.1=pypi_0
einops=0.8.0=pypi_0
evaluate=0.4.0=py39h06a4308_0
exceptiongroup=1.2.0=py39h06a4308_0
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.112.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py39h06a4308_0
flash-attn=2.6.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py39h5eee18b_0
fsspec=2024.2.0=pyhca7485f_0
gflags=2.2.2=he1b5a44_1004
glog=0.5.0=h48cff8f_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.48.2=he1ff14a_1
h11=0.14.0=py39h06a4308_0
httpcore=1.0.2=py39h06a4308_0
httptools=0.6.1=pypi_0
httpx=0.26.0=py39h06a4308_0
huggingface_hub=0.23.5=pyhd8ed1ab_0
icu=70.1=h27087fc_0
idna=3.7=py39h06a4308_0
importlib-metadata=8.2.0=pyha770c72_0
importlib_metadata=8.2.0=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
interegular=0.3.3=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.18.1=pyh707e725_3
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py39h06a4308_0
joblib=1.4.2=py39h06a4308_0
jpeg=9e=h5eee18b_1
jsonlines=2.0.0=pyhd3eb1b0_2
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py39hf3d152e_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lark=1.1.9=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.10.1.7=0
libcurand=10.3.6.82=0
libcurl=8.7.1=h251f7ec_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h516909a_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=14.1.0=h77fa898_0
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=14.1.0=h77fa898_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
lizard=1.17.10=pyhd8ed1ab_0
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.43.0=pypi_0
lm-format-enforcer=0.10.3=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown-it-py=2.2.0=py39h06a4308_1
markupsafe=2.1.3=py39h5eee18b_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.0=py39h06a4308_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
msgpack=1.0.8=pypi_0
multidict=6.0.4=py39h5eee18b_0
multiprocess=0.70.15=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py39h06a4308_0
ninja=1.11.1.1=pypi_0
ninja-base=1.10.2=hd09550d_5
nltk=3.8.1=py39h06a4308_0
numba=0.60.0=pypi_0
numexpr=2.8.7=py39h85018f9_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
nvidia-ml-py=12.555.43=pypi_0
openai=1.37.0=py39h06a4308_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_2
openssl=3.3.1=hb9d3cd8_3
orc=1.7.4=hb3bc3d3_1
outlines=0.0.46=pypi_0
packaging=24.1=py39h06a4308_0
pandas=2.2.2=py39h6a678d5_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=py39h5eee18b_0
pip=24.0=py39h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
prometheus-client=0.20.0=pypi_0
prometheus-fastapi-instrumentator=7.0.0=pypi_0
prompt-toolkit=3.0.47=pyha770c72_0
protobuf=5.27.3=pypi_0
psutil=6.0.0=py39hd3abc70_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
py-cpuinfo=9.0.0=pypi_0
pyairports=2.1.1=pypi_0
pyarrow=14.0.2=py39h1eedbd7_0
pycountry=24.6.1=pypi_0
pydantic=2.5.3=py39h06a4308_0
pydantic-core=2.14.6=py39hb02cf49_0
pygments=2.18.0=pyhd8ed1ab_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.19=h955ad1f_1
python-dateutil=2.9.0=pyhd8ed1ab_0
python-dotenv=1.0.1=pypi_0
python-slugify=5.0.2=pyhd3eb1b0_0
python-tzdata=2024.1=pyhd8ed1ab_0
python-xxhash=2.0.2=py39h5eee18b_1
python_abi=3.9=2_cp39
pytorch=2.4.0=py3.9_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py39h5eee18b_0
pyzmq=26.1.0=pypi_0
rank-bm25=0.2.2=pypi_0
ray=2.34.0=pypi_0
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2023.10.3=py39h5eee18b_0
requests=2.32.3=py39h06a4308_0
responses=0.13.3=pyhd3eb1b0_0
rich=13.7.1=py39h06a4308_0
rpds-py=0.20.0=pypi_0
s2n=1.3.27=hdbd6064_0
safetensors=0.4.2=py39ha89cbab_1
scikit-learn=1.5.1=py39h1128e8f_0
scipy=1.12.0=py39h5f9d8c6_0
sentencepiece=0.2.0=pypi_0
setuptools=69.5.1=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h6a678d5_1
sniffio=1.3.0=py39h06a4308_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.37.2=pypi_0
suffix-trees=0.3.0=pypi_0
sympy=1.12=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
text-unidecode=1.3=pyhd3eb1b0_0
threadpoolctl=3.5.0=py39h2f386ee_0
tiktoken=0.7.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=py39hff361bb_0
torchaudio=2.4.0=py39_cu118
torchtriton=3.0.0=py39
torchvision=0.19.0=py39_cu118
tornado=6.4.1=py39hd3abc70_0
tqdm=4.66.4=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.43.3=pyhd8ed1ab_0
typing-extensions=4.11.0=py39h06a4308_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024a=h04d1e81_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=2.2.2=py39h06a4308_0
utf8proc=2.6.1=h5eee18b_1
uvicorn=0.30.5=pypi_0
uvloop=0.19.0=pypi_0
vllm=0.5.4+cu118=pypi_0
vllm-flash-attn=2.6.1=pypi_0
watchfiles=0.23.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
websockets=12.0=pypi_0
wheel=0.43.0=py39h06a4308_0
xformers=0.0.27.post2=pypi_0
xxhash=0.8.0=h7f98852_3
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.7.2=py39hb9d737c_2
zeromq=4.3.5=h6a678d5_0
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
